pyproject.toml
[project]
name = "drone-control"
version = "0.1.0"
description = "Программа для управления дроном"
requires-python = ">=3.8"
dependencies = [
    "pymultiwii>=0.4.4",
    "pyserial==3.5",
    "python-dotenv==1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drone_control"]

[tool.rye]
managed = true
dev-dependencies = [] 
